requirements.txt
Django==3.1.7
vcrpy==4.1.1
requests==2.25.1